Intelligent Building Systems Improving Energy Performance Using Data Science and Analytics (revision 2-after submission date)
This file was added after the capstone submission date, and is not for grading purposes.  The intention of submitting this file is to ease the installation of required packages and libraries, using the "if(!require(PACKAGE_NAME))install.packages("PACKAGE_NAME", repos = "http://cran.us.r-project.org")".  Other minor changes have been made to codes (i.e. colors and tolerances), however the content remains the same.  Please review the "readme_new.txt" file to ensure that you are able to run the code, and download the additional file from Kaggle as necessary.